+++ commerce-domain/src/main/java/co/hooghly/commerce/repository/CustomerOptionRepository.java
@@ -14,9 +14,9 @@ public interface CustomerOptionRepository extends JpaRepository<CustomerOption, 
 	CustomerOption findOne(Long id);
 	
 	@Query("select o from CustomerOption o join fetch o.merchantStore om left join fetch o.descriptions od where om.id = ?1 and o.code = ?2")
-	CustomerOption findByCode(Integer merchantId, String code);
+	CustomerOption findByCode(Long merchantId, String code);
 	
 	@Query("select o from CustomerOption o join fetch o.merchantStore om left join fetch o.descriptions od where om.id = ?1 and od.language.id = ?2")
-	List<CustomerOption> findByStore(Integer merchantId, Integer languageId);
+	List<CustomerOption> findByStore(Long merchantId, Long languageId);
 
 }
