+++ commerce-domain/src/main/java/co/hooghly/commerce/repository/ProductOptionValueRepository.java
@@ -13,20 +13,20 @@ public interface ProductOptionValueRepository extends JpaRepository<ProductOptio
 	ProductOptionValue findOne(Long id);
 	
 	@Query("select p from ProductOptionValue p join fetch p.merchantStore pm left join fetch p.descriptions pd where p.id = ?2  and pm.id = ?1")
-	ProductOptionValue findOne(Integer storeId, Long id);
+	ProductOptionValue findOne(Long storeId, Long id);
 	
 	@Query("select distinct p from ProductOptionValue p join fetch p.merchantStore pm left join fetch p.descriptions pd where pm.id = ?1 and pd.language.id = ?2")
-	List<ProductOptionValue> findByStoreId(Integer storeId, Integer languageId);
+	List<ProductOptionValue> findByStoreId(Long storeId, Long languageId);
 	
 	@Query("select p from ProductOptionValue p join fetch p.merchantStore pm left join fetch p.descriptions pd where pm.id = ?1 and p.code = ?2")
-	public ProductOptionValue findByCode(Integer storeId, String optionValueCode);
+	public ProductOptionValue findByCode(Long storeId, String optionValueCode);
 	
 	@Query("select p from ProductOptionValue p join fetch p.merchantStore pm left join fetch p.descriptions pd where pm.id = ?1 and pd.name like %?2% and pd.language.id = ?3")
-	public List<ProductOptionValue> findByName(Integer storeId, String name, Integer languageId);
+	public List<ProductOptionValue> findByName(Long storeId, String name, Long languageId);
 	
 	
 	@Query("select distinct p from ProductOptionValue p join fetch p.merchantStore pm left join fetch p.descriptions pd where pm.id = ?1 and pd.language.id = ?2 and p.productOptionDisplayOnly = ?3")
-	public List<ProductOptionValue> findByReadOnly(Integer storeId, Integer languageId, boolean readOnly);
+	public List<ProductOptionValue> findByReadOnly(Long storeId, Long languageId, boolean readOnly);
 	
 
 }
