+++ commerce-domain/src/main/java/co/hooghly/commerce/repository/ShippingOriginRepository.java
@@ -8,5 +8,5 @@ import co.hooghly.commerce.domain.ShippingOrigin;
 public interface ShippingOriginRepository extends JpaRepository<ShippingOrigin, Long> {
 
 	@Query("select s from ShippingOrigin as s join fetch s.merchantStore sm where sm.id = ?1")
-	ShippingOrigin findByStore(Integer storeId);
+	ShippingOrigin findByStore(Long storeId);
 }
